requirements.txt
Flask
python-dotenv
psycopg2
SQLAlchemy
Flask-Migrate
Flask-SQLAlchemy
flask_wtf
pytest
openai
